requirements/base.txt
appdirs==1.4.4
attrs==21.2.0
backports.entry-points-selectable==1.1.0
black==21.7b0
certifi==2021.5.30
cfgv==3.3.0
chardet==4.0.0
click==8.0.1
distlib==0.3.2
filelock==3.0.12
Flask==2.0.1
identify==2.2.13
idna==2.10
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.0
pathspec==0.9.0
platformdirs==2.2.0
pluggy==0.13.1
pre-commit==2.14.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
PyYAML==5.4.1
regex==2021.8.3
requests==2.25.1
six==1.16.0
toml==0.10.2
tomli==1.2.1
tox==3.24.1
urllib3==1.26.6
virtualenv==20.7.2
Werkzeug==2.0.1
